refactor: use same api token form for creating integrations
