+++ web_src/src/components/IntegrationForm/ApiTokenForm.stories.tsx
@@ -1,9 +1,18 @@
-import React, { useState } from 'react'
+import React, { useState, useRef } from 'react'
 import type { Meta, StoryObj } from '@storybook/react'
 import { MemoryRouter } from 'react-router-dom'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { ApiTokenForm } from './ApiTokenForm'
 import type { IntegrationData, FormErrors } from './types'
 import { createMockSecrets, defaultProps } from '../../../test/__mocks__/secrets'
+
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: false,
+    },
+  },
+})
 
 const meta: Meta<typeof ApiTokenForm> = {
   title: 'Components/IntegrationForm/ApiTokenForm',
@@ -14,11 +23,13 @@ const meta: Meta<typeof ApiTokenForm> = {
   tags: ['autodocs'],
   decorators: [
     (Story) => (
-      <MemoryRouter>
-        <div className="w-[600px] p-6 bg-white dark:bg-zinc-900 rounded-lg">
-          <Story />
-        </div>
-      </MemoryRouter>
+      <QueryClientProvider client={queryClient}>
+        <MemoryRouter>
+          <div className="w-[600px] p-6 bg-white dark:bg-zinc-900 rounded-lg">
+            <Story />
+          </div>
+        </MemoryRouter>
+      </QueryClientProvider>
     ),
   ],
 }
@@ -29,7 +40,7 @@ type Story = StoryObj<typeof meta>
 const mockSecrets = createMockSecrets()
 
 export const NewSecretTab: Story = {
-  render: (args) => {
+  render: () => {
     const [integrationData, setIntegrationData] = useState<IntegrationData>({
       orgUrl: 'https://github.com/myorg',
       name: 'myorg-account',
@@ -42,6 +53,7 @@ export const NewSecretTab: Story = {
     const [errors, setErrors] = useState<FormErrors>({})
     const [apiTokenTab, setApiTokenTab] = useState<'existing' | 'new'>('new')
     const [newSecretToken, setNewSecretToken] = useState('')
+    const orgUrlRef = useRef<HTMLInputElement>(null)
 
     return (
       <ApiTokenForm
@@ -53,7 +65,7 @@ export const NewSecretTab: Story = {
         setApiTokenTab={setApiTokenTab}
         newSecretToken={newSecretToken}
         setNewSecretToken={setNewSecretToken}
-        {...args}
+        orgUrlRef={orgUrlRef}
         secrets={mockSecrets}
         organizationId={defaultProps.organizationId}
         canvasId={defaultProps.canvasId}
@@ -63,7 +75,7 @@ export const NewSecretTab: Story = {
 }
 
 export const ExistingSecretTab: Story = {
-  render: (args) => {
+  render: () => {
     const [integrationData, setIntegrationData] = useState<IntegrationData>({
       orgUrl: 'https://github.com/myorg',
       name: 'myorg-account',
@@ -76,6 +88,7 @@ export const ExistingSecretTab: Story = {
     const [errors, setErrors] = useState<FormErrors>({})
     const [apiTokenTab, setApiTokenTab] = useState<'existing' | 'new'>('existing')
     const [newSecretToken, setNewSecretToken] = useState('')
+    const orgUrlRef = useRef<HTMLInputElement>(null)
 
     return (
       <ApiTokenForm
@@ -87,7 +100,7 @@ export const ExistingSecretTab: Story = {
         setApiTokenTab={setApiTokenTab}
         newSecretToken={newSecretToken}
         setNewSecretToken={setNewSecretToken}
-        {...args}
+        orgUrlRef={orgUrlRef}
         secrets={mockSecrets}
         organizationId={defaultProps.organizationId}
         canvasId={defaultProps.canvasId}
@@ -97,7 +110,7 @@ export const ExistingSecretTab: Story = {
 }
 
 export const ExistingSecretSelected: Story = {
-  render: (args) => {
+  render: () => {
     const [integrationData, setIntegrationData] = useState<IntegrationData>({
       orgUrl: 'https://github.com/myorg',
       name: 'myorg-account',
@@ -110,6 +123,7 @@ export const ExistingSecretSelected: Story = {
     const [errors, setErrors] = useState<FormErrors>({})
     const [apiTokenTab, setApiTokenTab] = useState<'existing' | 'new'>('existing')
     const [newSecretToken, setNewSecretToken] = useState('')
+    const orgUrlRef = useRef<HTMLInputElement>(null)
 
     return (
       <ApiTokenForm
@@ -121,7 +135,7 @@ export const ExistingSecretSelected: Story = {
         setApiTokenTab={setApiTokenTab}
         newSecretToken={newSecretToken}
         setNewSecretToken={setNewSecretToken}
-        {...args}
+        orgUrlRef={orgUrlRef}
         secrets={mockSecrets}
         organizationId={defaultProps.organizationId}
         canvasId={defaultProps.canvasId}
@@ -131,7 +145,7 @@ export const ExistingSecretSelected: Story = {
 }
 
 export const WithNewSecretValue: Story = {
-  render: (args) => {
+  render: () => {
     const [integrationData, setIntegrationData] = useState<IntegrationData>({
       orgUrl: 'https://api.semaphoreci.com',
       name: 'semaphore-integration',
@@ -144,6 +158,7 @@ export const WithNewSecretValue: Story = {
     const [errors, setErrors] = useState<FormErrors>({})
     const [apiTokenTab, setApiTokenTab] = useState<'existing' | 'new'>('new')
     const [newSecretToken, setNewSecretToken] = useState('smp_1234567890abcdef1234567890abcdef')
+    const orgUrlRef = useRef<HTMLInputElement>(null)
 
     return (
       <ApiTokenForm
@@ -155,7 +170,7 @@ export const WithNewSecretValue: Story = {
         setApiTokenTab={setApiTokenTab}
         newSecretToken={newSecretToken}
         setNewSecretToken={setNewSecretToken}
-        {...args}
+        orgUrlRef={orgUrlRef}
         secrets={mockSecrets}
         organizationId={defaultProps.organizationId}
         canvasId={defaultProps.canvasId}
@@ -165,7 +180,7 @@ export const WithNewSecretValue: Story = {
 }
 
 export const WithErrors: Story = {
-  render: (args) => {
+  render: () => {
     const [integrationData, setIntegrationData] = useState<IntegrationData>({
       orgUrl: 'https://github.com/myorg',
       name: 'myorg-account',
@@ -181,6 +196,7 @@ export const WithErrors: Story = {
     })
     const [apiTokenTab, setApiTokenTab] = useState<'existing' | 'new'>('new')
     const [newSecretToken, setNewSecretToken] = useState('')
+    const orgUrlRef = useRef<HTMLInputElement>(null)
 
     return (
       <ApiTokenForm
@@ -192,7 +208,7 @@ export const WithErrors: Story = {
         setApiTokenTab={setApiTokenTab}
         newSecretToken={newSecretToken}
         setNewSecretToken={setNewSecretToken}
-        {...args}
+        orgUrlRef={orgUrlRef}
         secrets={mockSecrets}
         organizationId={defaultProps.organizationId}
         canvasId={defaultProps.canvasId}
@@ -202,7 +218,7 @@ export const WithErrors: Story = {
 }
 
 export const NoSecretsAvailable: Story = {
-  render: (args) => {
+  render: () => {
     const [integrationData, setIntegrationData] = useState<IntegrationData>({
       orgUrl: 'https://github.com/myorg',
       name: 'myorg-account',
@@ -215,6 +231,7 @@ export const NoSecretsAvailable: Story = {
     const [errors, setErrors] = useState<FormErrors>({})
     const [apiTokenTab, setApiTokenTab] = useState<'existing' | 'new'>('existing')
     const [newSecretToken, setNewSecretToken] = useState('')
+    const orgUrlRef = useRef<HTMLInputElement>(null)
 
     return (
       <ApiTokenForm
@@ -226,17 +243,17 @@ export const NoSecretsAvailable: Story = {
         setApiTokenTab={setApiTokenTab}
         newSecretToken={newSecretToken}
         setNewSecretToken={setNewSecretToken}
+        orgUrlRef={orgUrlRef}
         secrets={[]}
         organizationId="org-123"
         canvasId="canvas-456"
-        {...args}
       />
     )
   }
 }
 
 export const MultipleKeyOptions: Story = {
-  render: (args) => {
+  render: () => {
     const [integrationData, setIntegrationData] = useState<IntegrationData>({
       orgUrl: 'https://api.example.com',
       name: 'multi-key-integration',
@@ -249,6 +266,7 @@ export const MultipleKeyOptions: Story = {
     const [errors, setErrors] = useState<FormErrors>({})
     const [apiTokenTab, setApiTokenTab] = useState<'existing' | 'new'>('existing')
     const [newSecretToken, setNewSecretToken] = useState('')
+    const orgUrlRef = useRef<HTMLInputElement>(null)
 
     return (
       <ApiTokenForm
@@ -260,7 +278,46 @@ export const MultipleKeyOptions: Story = {
         setApiTokenTab={setApiTokenTab}
         newSecretToken={newSecretToken}
         setNewSecretToken={setNewSecretToken}
-        {...args}
+        orgUrlRef={orgUrlRef}
+        secrets={mockSecrets}
+        organizationId={defaultProps.organizationId}
+        canvasId={defaultProps.canvasId}
+      />
+    )
+  }
+}
+
+export const EditMode: Story = {
+  render: () => {
+    const [integrationData, setIntegrationData] = useState<IntegrationData>({
+      orgUrl: 'https://api.semaphoreci.com',
+      name: 'semaphore-integration',
+      apiToken: {
+        secretName: 'semaphore-api-key',
+        secretKey: 'api-token'
+      }
+    })
+
+    const [errors, setErrors] = useState<FormErrors>({})
+    const [apiTokenTab, setApiTokenTab] = useState<'existing' | 'new'>('existing')
+    const [newSecretToken, setNewSecretToken] = useState('')
+    const [newSecretValue, setNewSecretValue] = useState('')
+    const orgUrlRef = useRef<HTMLInputElement>(null)
+
+    return (
+      <ApiTokenForm
+        integrationData={integrationData}
+        setIntegrationData={setIntegrationData}
+        errors={errors}
+        setErrors={setErrors}
+        apiTokenTab={apiTokenTab}
+        setApiTokenTab={setApiTokenTab}
+        newSecretToken={newSecretToken}
+        setNewSecretToken={setNewSecretToken}
+        orgUrlRef={orgUrlRef}
+        isEditMode={true}
+        newSecretValue={newSecretValue}
+        setNewSecretValue={setNewSecretValue}
         secrets={mockSecrets}
         organizationId={defaultProps.organizationId}
         canvasId={defaultProps.canvasId}
