+++ web_src/src/components/IntegrationForm/index.ts
@@ -1,7 +1,6 @@
 export { GitHubIntegrationForm } from './GitHubIntegrationForm';
 export { SemaphoreIntegrationForm } from './SemaphoreIntegrationForm';
 export { ApiTokenForm } from './ApiTokenForm';
-export { SettingsApiTokenForm } from './SettingsApiTokenForm';
 export { useIntegrationForm, NEW_SECRET_NAME } from './useIntegrationForm';
 export { getIntegrationConfig } from './integrationConfigs';
 export type { IntegrationData, FormErrors, BaseIntegrationFormProps, IntegrationConfig } from './types';
